Reordered randomize_state arguments and changed bitfield type to size_t
* Always put bond_lim as last argument

+++ source/algorithms/AlgorithmFinite.h
@@ -47,7 +47,7 @@ class AlgorithmFinite : public AlgorithmBase {
     void         clear_convergence_status() override;
     void         randomize_state(ResetReason reason, StateInit state_init, std::optional<StateInitType> state_type = std::nullopt,
                                  std::optional<std::string> sector = std::nullopt, std::optional<bool> use_eigenspinors = std::nullopt,
-                                 std::optional<long> bitfield = std::nullopt, std::optional<long> bond_lim = std::nullopt,
+                                 std::optional<size_t> bitfield = std::nullopt, std::optional<long> bond_lim = std::nullopt,
                                  std::optional<double> trnc_lim = std::nullopt);
 
     void write_to_file(StorageReason storage_reason = StorageReason::CHECKPOINT, std::optional<CopyPolicy> copy_file = std::nullopt) override;
